Desktop/linkedincontent/requirements.txt
streamlit>=1.51.0
requests>=2.32.0
pdfplumber>=0.11.0
python-docx>=1.2.0
pandas>=2.3.0

